Add isCompressed to PutRequest.
A new version of PutRequest is introduced but disabled.
The new version has aded the "isCompressed" field.
